refactor: 4단계 - 수동 로또 개수 유효성 체크 로직 이동 및 LottoTicketPrice > LottoTicketValidator 파일명 수정
