+++ src/test/java/lotto/LottoTicketPriceTest.java
@@ -10,7 +10,7 @@ public class LottoTicketPriceTest {
     @Test
     @DisplayName("구입금액에 맞게 구매한 로또 개수 출력")
     void 로또_구매() {
-        assertThat(LottoTicketPrice.countByAmount(14000)).isEqualTo(14);
+        assertThat(LottoTicketValidator.countByAmount(14000)).isEqualTo(14);
     }
 
     @Test
@@ -18,7 +18,7 @@ public class LottoTicketPriceTest {
     void 로또_구매_적은_금액() {
         assertThatExceptionOfType(IllegalArgumentException.class)
                 .isThrownBy(() -> {
-                    LottoTicketPrice.countByAmount(900);
+                    LottoTicketValidator.countByAmount(900);
                 }).withMessageMatching("지불한 금액이 로또 금액보다 적습니다.");
     }
 }
